refactor: 프로젝트 구조 리팩토링 #9
