+++ App.js
@@ -5,34 +5,34 @@ import { PersistGate } from "redux-persist/integration/react";
 import { store, persistor } from "./store/store";
 import { NavigationContainer } from "@react-navigation/native";
 import { createStackNavigator } from "@react-navigation/stack";
-import StartScreen from "./screens/startscreen";
-import LoginScreen from "./screens/login/LoginScreen";
-import SignUpScreen from "./screens/signup/SignupScreen";
-import SignupScreenEmail from "./screens/signup/SignupScreenEmail";
-import SignupScreenPW from "./screens/signup/SignupScreenPW";
-import SignupScreenPWCheck from "./screens/signup/SignupScreenPWCheck";
-import SignupOK from "./screens/signup/SignupOK";
-import MainPage from "./MainPage";
-import MakeDiary from "./MakeDiary";
-import MakeDiaryAI from "./MakeDiaryAI";
-import ChooseDrawStyle from "./ChooseDrawStyle";
-import ChooseDate from "./ChooseDate";
-import WriteDiary from "./WriteDiary";
-import ChooseWeather from "./ChooseWeather";
-import ChooseDetails from "./ChooseDetails";
-import AIHelpDiary from "./AIHelpDiary";
-import AIWriteDiary from "./AIWriteDiary";
-import AICompleteDiary from "./AICompleteDiary";
-import MyPage from "./screens/mypage/MyPage";
-import UserInfo from "./UserInfo";
-import EditName from "./EditName";
-import Announcements from "./Announcements";
-import ChangePasswordStep1 from "./ChangePasswordStep1";
-import ChangePasswordStep2 from "./ChangePasswordStep2";
-import ChangePasswordStep3 from "./ChangePasswordStep3";
-import ShowDiary from "./ShowDiary";
-import SignupScreenValidationEmail from "./screens/signup/SignupScreenValidationEmail";
-import ResetPassword from "./screens/find/ResetPassword";
+import StartScreen from "./screens/startScreen";
+import LoginScreen from "./screens/auth/login/LoginScreen";
+import SignUpScreen from "./screens/auth/signup/SignupScreen";
+import SignupScreenEmail from "./screens/auth/signup/SignupScreenEmail";
+import SignupScreenPW from "./screens/auth/signup/SignupScreenPW";
+import SignupScreenPWCheck from "./screens/auth/signup/SignupScreenPWCheck";
+import SignupOK from "./screens/auth/signup/SignupOK";
+import MainPage from "./screens/MainPage";
+import MakeDiary from "./screens/diary/MakeDiary";
+import MakeDiaryAI from "./screens/diary/MakeDiaryAI";
+import ChooseDrawStyle from "./screens/diary/ChooseDrawStyle";
+import ChooseDate from "./screens/diary/ChooseDate";
+import WriteDiary from "./screens/diary/WriteDiary";
+import ChooseWeather from "./screens/diary/ChooseWeather";
+import ChooseDetails from "./screens/diary/ChooseDetails";
+import AIHelpDiary from "./screens/diary/AIHelpDiary";
+import AIWriteDiary from "./screens/diary/AIWriteDiary";
+import AICompleteDiary from "./screens/diary/AICompleteDiary";
+import MyPage from "./screens/user/MyPage";
+import UserInfo from "./screens/user/UserInfo";
+import EditName from "./screens/user/EditName";
+import Announcements from "./screens/Announcements";
+import ChangePasswordStep1 from "./screens/auth/password/ChangePasswordStep1";
+import ChangePasswordStep2 from "./screens/auth/password/ChangePasswordStep2";
+import ChangePasswordStep3 from "./screens/auth/password/ChangePasswordStep3";
+import ShowDiary from "./screens/diary/ShowDiary";
+import SignupScreenValidationEmail from "./screens/auth/signup/SignupScreenValidationEmail";
+import ResetPassword from "./screens/auth/password/ResetPassword";
 import { Use } from "react-native-svg";
 
 const Stack = createStackNavigator();
